rename file with appender interface
